pyproject.toml
[project]
name = "sharp-fhir-mcp"
version = "0.1.0"
description = "SHARP-on-MCP compliant FHIR R4 MCP server with MCP-UI clinical dashboards"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "SHARP-FHIR-MCP contributors" },
]
keywords = [
    "mcp",
    "model-context-protocol",
    "sharp",
    "fhir",
    "healthcare",
    "ehr",
    "clinical",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "mcp>=1.0.0",
    "mcp-ui>=0.1.2",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[project.scripts]
sharp-fhir-mcp = "sharp_fhir_mcp.server:main"

[project.urls]
Homepage = "https://www.sharponmcp.com"
Specification = "https://www.sharponmcp.com/overview.html"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sharp_fhir_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
